requirements.txt
gym>=0.18.3
opencv-python>=4.1.0.25
setuptools>=49.2.0
tqdm>=4.32.2
numpy>=1.16.2
requests>=2.20.0
ipython>=7.5.0
typing>=3.6.6
lxml>=4.3.3
psutil>=5.6.2
Pyro4>=4.76
getch>=1.0; sys_platform != 'win32' and sys_platform != 'cygwin'
coloredlogs>=10.0
pillow>=8.0.0,<=8.2.0
matplotlib>=3.2.2
dill>=0.3.1.1
daemoniker>=0.2.3
xmltodict==0.12.0
inflection>=0.3.1
jinja2>=2.11.2
simple-term-menu
bullet
minerl-patched==0.1.0
pandas==1.1.3
azure-storage-blob==12.9.0
